refactor: move sanitizedDomain function
