+++ cmd/internal/storage/certificates.go
@@ -4,6 +4,10 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
+	"strings"
+
+	"github.com/go-acme/lego/v5/log"
+	"golang.org/x/net/idna"
 )
 
 const (
@@ -55,4 +59,17 @@ func getCertificatesRootPath(basePath string) string {
 
 func getCertificatesArchivePath(basePath string) string {
 	return filepath.Join(basePath, baseArchivesFolderName)
+}
+
+// sanitizedDomain Make sure no funny chars are in the cert names (like wildcards ;)).
+func sanitizedDomain(domain string) string {
+	safe, err := idna.ToASCII(strings.NewReplacer(":", "-", "*", "_").Replace(domain))
+	if err != nil {
+		log.Fatal("Could not sanitize the domain.",
+			log.DomainAttr(domain),
+			log.ErrorAttr(err),
+		)
+	}
+
+	return safe
 }
